Fixed layout and bugs
